+++ src/datacamp_downloader/templates/course.py
